+++ AssetsFlow.Client.React/src/Components/Journal/DataElementGroups/EntriesList.jsx
@@ -1,16 +1,12 @@
-import React from 'react';
-import DataElement from '../DataElement/DataElement';
+import React from "react";
+import DataElement from "../DataElement/DataElement";
 
-function EntriesList({ entries, processCellUpdate, overviewType }) {
+function EntriesList({ entries, overviewType }) {
   return (
     <ul className="flex flex-wrap gap-1">
       {entries.map((entry) => (
-        <li key={entry.id} className='p-1'>
-          <DataElement
-            cellInfo={entry}
-            onCellUpdate={processCellUpdate}
-            overviewType={overviewType}
-          />
+        <li key={entry.id} className="p-1">
+          <DataElement cellInfo={entry} overviewType={overviewType} />
         </li>
       ))}
     </ul>