+++ AssetsFlow.Client.React/src/Components/Journal/TradeComposite/Controls/CompositeControls.jsx
@@ -1,35 +1,23 @@
 import React, { useState } from "react";
 import * as Constants from "@constants/journalConstants";
-import { ProcessingStatus } from "@constants/constants";
 import ProcessingAndSuccessMessage from "@components/Processing/ProcessingAndSuccessMessage";
 import ActionButtons from "./ActionButtons";
 import TradeClosingModal from "./TradeClosingModal";
 import { useTradeActionMutation } from "@hooks/Journal/Composite/useTradeActionMutation";
-import { useProcessingWrapper } from "@hooks/useProcessingWrapper";
 
-function CompositeControls({ tradeComposite, onTradeActionSuccess }) {
-  const { processingStatus, setNewStatus } = useProcessingWrapper(ProcessingStatus.NONE);
+function CompositeControls({ tradeComposite }) {
   const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const tradeActionMutation = useTradeActionMutation(
-    tradeComposite,
-    (response) => {
-      onTradeActionSuccess(response);
-      setNewStatus(ProcessingStatus.SUCCESS);
-    }
-  );
+  const { tradeActionMutation, processingStatus } = useTradeActionMutation(tradeComposite);
 
   const handleAction = (action) => {
     if (action === Constants.TradeActions.CLOSE) {
       setIsModalOpen(true);
     } else {
-      setNewStatus(ProcessingStatus.PROCESSING);
       tradeActionMutation.mutate({ action, additionalParam: null });
     }
   };
 
   const handleCloseTrade = (closingPrice) => {
-    setNewStatus(ProcessingStatus.PROCESSING);
     tradeActionMutation.mutate({ action: Constants.TradeActions.CLOSE, additionalParam: closingPrice });
     setIsModalOpen(false);
   };