+++ AssetsFlow.Client.React/src/Hooks/Journal/Element/useCacheNewElement.jsx
@@ -15,5 +15,5 @@ export function useCacheNewElement(tradeComposite) {
     );
   };
 
-  return { onElementUpdate };
+  return onElementUpdate;
 }