+++ AssetsFlow.Client.React/src/Hooks/Journal/Element/useElementActionMutation.jsx
@@ -3,9 +3,17 @@ import { ElementActions } from "@constants/journalConstants";
 import { ProcessingStatus } from "@constants/constants";
 import { timestampElementAPI, deleteElementAPI } from "@services/ApiRequests/elementApiAccess";
 import { useProcessingWrapper } from "@hooks/useProcessingWrapper";
+import { useUpdateTradeStatusFromResponse } from "../Composite/useUpdateTradeStatusFromResponse";
+import { useUpdateElementCacheData } from "@hooks/Journal/Element/useUpdateElementCacheData";
+import { newStatesResponseParser } from "@services/newStatesResponseParser";
+import { useRemoveElementFromTrade } from "@hooks/Journal/Element/useRemoveElementFromTrade"
+import * as Constants from "@constants/journalConstants";
 
-export function useElementActionMutation(tradeElement, onActionSuccess) {
+export function useElementActionMutation(tradeElement) {
   const { processingStatus, setNewStatus } = useProcessingWrapper(ProcessingStatus.NONE);
+  const updateTradeStatuses = useUpdateTradeStatusFromResponse(tradeElement[Constants.ELEMENT_COMPOSITEFK_STING]);
+  const updateElementProp = useUpdateElementCacheData(tradeElement[Constants.ELEMENT_COMPOSITEFK_STING], tradeElement.id);
+    const removeElement = useRemoveElementFromTrade(tradeElement[Constants.ELEMENT_COMPOSITEFK_STING]);
 
   const elementActionMutation = useMutation({
     mutationFn: async ({ action }) => {
@@ -30,9 +38,15 @@ export function useElementActionMutation(tradeElement, onActionSuccess) {
       console.error("Error performing action:", error);
     },
     onSuccess: ({ response }, { action }) => {
-      if (onActionSuccess) {
-        onActionSuccess(action, response);
+          if (action === ElementActions.DELETE) {
+      removeElement(tradeElement.id);
+    }
+      // Update timestamp if present
+      const { elementsNewTimeStamp } = newStatesResponseParser(response);
+      if (elementsNewTimeStamp) {
+        updateElementProp(Constants.ELEMENT_TIMESTAMP_STING, elementsNewTimeStamp);
       }
+      updateTradeStatuses(response);
       setNewStatus(ProcessingStatus.SUCCESS); // Set to SUCCESS when mutation is successful
     },
   });