+++ AssetsFlow.Client.React/src/Hooks/Journal/Element/useRemoveElementFromTrade.jsx
@@ -5,16 +5,18 @@ import { tradeKeysFactory } from "@services/query-key-factory";
 
 export function useRemoveElementFromTrade(tradeId) {
   const queryClient = useQueryClient();
-  const tradeAndIdArray = tradeKeysFactory.tradeAndIdArrayKey(tradeId)
+  const tradeAndIdArray = tradeKeysFactory.tradeAndIdArrayKey(tradeId);
 
   const removeElement = (elementId) => {
     queryClient.setQueryData(tradeAndIdArray, (oldTrade) => {
       if (!oldTrade) return oldTrade;
       return produce(oldTrade, (draft) => {
-        draft.tradeElements = draft.tradeElements.filter(ele => ele.id !== elementId);
+        draft.tradeElements = draft.tradeElements.filter(
+          (ele) => ele.id !== elementId
+        );
       });
     });
   };
 
-  return { removeElement };
+  return removeElement;
 }