+++ AssetsFlow.Server/CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.9.503-s] - 2025-06-..
 - IsAnyContentMissing prop on elementDTO
 - Reworked contentEdit logic - always load trade for now, for uptodate ismissing data (to rethink)
+- CompositeFK name mismatch (entity and dto)
 
 
 ## [0.9.502-s] - 2025-06-22